Probability Reference Sample 1 with Bootstrap Replicate Weights (sp1_bootstrap)

A replicate-weight version of sp1, including the main survey weight and 500 bootstrap replicate weights. It is provided to illustrate design-based variance estimation with svrepdesign. The original primary sampling unit and stratum identifiers are not included.

Usage

data(sp1_bootstrap)

Format

A data frame with 3494 rows and 512 columns. The first 12 columns are the substantive survey variables from sp1 (agecat, marital, race, education, employment, smoking, comorbidity, psa_level, BMI, diabetes, pros_enlarged, wts_sp1). The remaining 500 columns are bootstrap replicate weights named bw1 through bw500 (numeric).

Source

Derived from sp1 (National Health and Nutrition Examination Survey, NHANES, 1999–2010 cycles), with bootstrap replicate weights added by the package authors using the Rao-Wu rescaling bootstrap.

Details

The bootstrap replicate weights were constructed from the original stratified cluster design of sp1, using the Rao-Wu rescaling bootstrap method. A total of R = 500 replicates were produced with seed = 2026.

The variables psu_sp1 and strata_sp1 are not included in this dataset because they are not needed when using replicate weights for variance estimation. The main survey weight wts_sp1 and the replicate weight columns bw1–bw500 are sufficient for constructing a replicate-weight survey design object via survey::svrepdesign().

Examples

data(sp1_bootstrap)

# Example: create replicate-weight survey design object
des_boot <- survey::svrepdesign(
  data    = sp1_bootstrap,
  weights = ~wts_sp1,
  repweights = "bw[0-9]+",
  type    = "bootstrap",
  combined.weights = FALSE
)
#> Warning: Data look like combined weights: mean replication weight is 5950.55221390828  and mean sampling weight is 5952.20716602519

summary(des_boot)
#> Call: svrepdesign.default(data = sp1_bootstrap, weights = ~wts_sp1, 
#>     repweights = "bw[0-9]+", type = "bootstrap", combined.weights = FALSE)
#> Survey bootstrap with 500 replicates.
#> Variables: 
#>   [1] "agecat"        "marital"       "race"          "education"    
#>   [5] "employment"    "smoking"       "comorbidity"   "psa_level"    
#>   [9] "BMI"           "diabetes"      "pros_enlarged" "wts_sp1"      
#>  [13] "bw1"           "bw2"           "bw3"           "bw4"          
#>  [17] "bw5"           "bw6"           "bw7"           "bw8"          
#>  [21] "bw9"           "bw10"          "bw11"          "bw12"         
#>  [25] "bw13"          "bw14"          "bw15"          "bw16"         
#>  [29] "bw17"          "bw18"          "bw19"          "bw20"         
#>  [33] "bw21"          "bw22"          "bw23"          "bw24"         
#>  [37] "bw25"          "bw26"          "bw27"          "bw28"         
#>  [41] "bw29"          "bw30"          "bw31"          "bw32"         
#>  [45] "bw33"          "bw34"          "bw35"          "bw36"         
#>  [49] "bw37"          "bw38"          "bw39"          "bw40"         
#>  [53] "bw41"          "bw42"          "bw43"          "bw44"         
#>  [57] "bw45"          "bw46"          "bw47"          "bw48"         
#>  [61] "bw49"          "bw50"          "bw51"          "bw52"         
#>  [65] "bw53"          "bw54"          "bw55"          "bw56"         
#>  [69] "bw57"          "bw58"          "bw59"          "bw60"         
#>  [73] "bw61"          "bw62"          "bw63"          "bw64"         
#>  [77] "bw65"          "bw66"          "bw67"          "bw68"         
#>  [81] "bw69"          "bw70"          "bw71"          "bw72"         
#>  [85] "bw73"          "bw74"          "bw75"          "bw76"         
#>  [89] "bw77"          "bw78"          "bw79"          "bw80"         
#>  [93] "bw81"          "bw82"          "bw83"          "bw84"         
#>  [97] "bw85"          "bw86"          "bw87"          "bw88"         
#> [101] "bw89"          "bw90"          "bw91"          "bw92"         
#> [105] "bw93"          "bw94"          "bw95"          "bw96"         
#> [109] "bw97"          "bw98"          "bw99"          "bw100"        
#> [113] "bw101"         "bw102"         "bw103"         "bw104"        
#> [117] "bw105"         "bw106"         "bw107"         "bw108"        
#> [121] "bw109"         "bw110"         "bw111"         "bw112"        
#> [125] "bw113"         "bw114"         "bw115"         "bw116"        
#> [129] "bw117"         "bw118"         "bw119"         "bw120"        
#> [133] "bw121"         "bw122"         "bw123"         "bw124"        
#> [137] "bw125"         "bw126"         "bw127"         "bw128"        
#> [141] "bw129"         "bw130"         "bw131"         "bw132"        
#> [145] "bw133"         "bw134"         "bw135"         "bw136"        
#> [149] "bw137"         "bw138"         "bw139"         "bw140"        
#> [153] "bw141"         "bw142"         "bw143"         "bw144"        
#> [157] "bw145"         "bw146"         "bw147"         "bw148"        
#> [161] "bw149"         "bw150"         "bw151"         "bw152"        
#> [165] "bw153"         "bw154"         "bw155"         "bw156"        
#> [169] "bw157"         "bw158"         "bw159"         "bw160"        
#> [173] "bw161"         "bw162"         "bw163"         "bw164"        
#> [177] "bw165"         "bw166"         "bw167"         "bw168"        
#> [181] "bw169"         "bw170"         "bw171"         "bw172"        
#> [185] "bw173"         "bw174"         "bw175"         "bw176"        
#> [189] "bw177"         "bw178"         "bw179"         "bw180"        
#> [193] "bw181"         "bw182"         "bw183"         "bw184"        
#> [197] "bw185"         "bw186"         "bw187"         "bw188"        
#> [201] "bw189"         "bw190"         "bw191"         "bw192"        
#> [205] "bw193"         "bw194"         "bw195"         "bw196"        
#> [209] "bw197"         "bw198"         "bw199"         "bw200"        
#> [213] "bw201"         "bw202"         "bw203"         "bw204"        
#> [217] "bw205"         "bw206"         "bw207"         "bw208"        
#> [221] "bw209"         "bw210"         "bw211"         "bw212"        
#> [225] "bw213"         "bw214"         "bw215"         "bw216"        
#> [229] "bw217"         "bw218"         "bw219"         "bw220"        
#> [233] "bw221"         "bw222"         "bw223"         "bw224"        
#> [237] "bw225"         "bw226"         "bw227"         "bw228"        
#> [241] "bw229"         "bw230"         "bw231"         "bw232"        
#> [245] "bw233"         "bw234"         "bw235"         "bw236"        
#> [249] "bw237"         "bw238"         "bw239"         "bw240"        
#> [253] "bw241"         "bw242"         "bw243"         "bw244"        
#> [257] "bw245"         "bw246"         "bw247"         "bw248"        
#> [261] "bw249"         "bw250"         "bw251"         "bw252"        
#> [265] "bw253"         "bw254"         "bw255"         "bw256"        
#> [269] "bw257"         "bw258"         "bw259"         "bw260"        
#> [273] "bw261"         "bw262"         "bw263"         "bw264"        
#> [277] "bw265"         "bw266"         "bw267"         "bw268"        
#> [281] "bw269"         "bw270"         "bw271"         "bw272"        
#> [285] "bw273"         "bw274"         "bw275"         "bw276"        
#> [289] "bw277"         "bw278"         "bw279"         "bw280"        
#> [293] "bw281"         "bw282"         "bw283"         "bw284"        
#> [297] "bw285"         "bw286"         "bw287"         "bw288"        
#> [301] "bw289"         "bw290"         "bw291"         "bw292"        
#> [305] "bw293"         "bw294"         "bw295"         "bw296"        
#> [309] "bw297"         "bw298"         "bw299"         "bw300"        
#> [313] "bw301"         "bw302"         "bw303"         "bw304"        
#> [317] "bw305"         "bw306"         "bw307"         "bw308"        
#> [321] "bw309"         "bw310"         "bw311"         "bw312"        
#> [325] "bw313"         "bw314"         "bw315"         "bw316"        
#> [329] "bw317"         "bw318"         "bw319"         "bw320"        
#> [333] "bw321"         "bw322"         "bw323"         "bw324"        
#> [337] "bw325"         "bw326"         "bw327"         "bw328"        
#> [341] "bw329"         "bw330"         "bw331"         "bw332"        
#> [345] "bw333"         "bw334"         "bw335"         "bw336"        
#> [349] "bw337"         "bw338"         "bw339"         "bw340"        
#> [353] "bw341"         "bw342"         "bw343"         "bw344"        
#> [357] "bw345"         "bw346"         "bw347"         "bw348"        
#> [361] "bw349"         "bw350"         "bw351"         "bw352"        
#> [365] "bw353"         "bw354"         "bw355"         "bw356"        
#> [369] "bw357"         "bw358"         "bw359"         "bw360"        
#> [373] "bw361"         "bw362"         "bw363"         "bw364"        
#> [377] "bw365"         "bw366"         "bw367"         "bw368"        
#> [381] "bw369"         "bw370"         "bw371"         "bw372"        
#> [385] "bw373"         "bw374"         "bw375"         "bw376"        
#> [389] "bw377"         "bw378"         "bw379"         "bw380"        
#> [393] "bw381"         "bw382"         "bw383"         "bw384"        
#> [397] "bw385"         "bw386"         "bw387"         "bw388"        
#> [401] "bw389"         "bw390"         "bw391"         "bw392"        
#> [405] "bw393"         "bw394"         "bw395"         "bw396"        
#> [409] "bw397"         "bw398"         "bw399"         "bw400"        
#> [413] "bw401"         "bw402"         "bw403"         "bw404"        
#> [417] "bw405"         "bw406"         "bw407"         "bw408"        
#> [421] "bw409"         "bw410"         "bw411"         "bw412"        
#> [425] "bw413"         "bw414"         "bw415"         "bw416"        
#> [429] "bw417"         "bw418"         "bw419"         "bw420"        
#> [433] "bw421"         "bw422"         "bw423"         "bw424"        
#> [437] "bw425"         "bw426"         "bw427"         "bw428"        
#> [441] "bw429"         "bw430"         "bw431"         "bw432"        
#> [445] "bw433"         "bw434"         "bw435"         "bw436"        
#> [449] "bw437"         "bw438"         "bw439"         "bw440"        
#> [453] "bw441"         "bw442"         "bw443"         "bw444"        
#> [457] "bw445"         "bw446"         "bw447"         "bw448"        
#> [461] "bw449"         "bw450"         "bw451"         "bw452"        
#> [465] "bw453"         "bw454"         "bw455"         "bw456"        
#> [469] "bw457"         "bw458"         "bw459"         "bw460"        
#> [473] "bw461"         "bw462"         "bw463"         "bw464"        
#> [477] "bw465"         "bw466"         "bw467"         "bw468"        
#> [481] "bw469"         "bw470"         "bw471"         "bw472"        
#> [485] "bw473"         "bw474"         "bw475"         "bw476"        
#> [489] "bw477"         "bw478"         "bw479"         "bw480"        
#> [493] "bw481"         "bw482"         "bw483"         "bw484"        
#> [497] "bw485"         "bw486"         "bw487"         "bw488"        
#> [501] "bw489"         "bw490"         "bw491"         "bw492"        
#> [505] "bw493"         "bw494"         "bw495"         "bw496"        
#> [509] "bw497"         "bw498"         "bw499"         "bw500"